requirements.txt
        Flask
        Flask-Cors
        gunicorn
        Flask-SQLAlchemy
        psycopg2-binary
        ibm-watson
        python-dotenv
        flask-limiter
        flask-swagger-ui